feat(maestro): wip adjusting to each event and fixing collector creation/update on Sinks, only in sinker update.
